requirements.txt
# GPS Position System Requirements
# Python 3.8+

# Core Data Processing
numpy==1.24.3
scipy==1.10.1
pandas==2.0.3

# GUI & Visualization
PyQt6==6.5.2
pyqtgraph==0.13.3
Pillow==10.0.0

# Network I/O
requests==2.31.0

# NMEA Parsing
pynmea2==1.19.0

# CLI Tools
click==8.1.7

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1

# Code Quality & Formatting
black==23.7.0
flake8==6.1.0
mypy==1.5.1
pylint==2.17.5

# Development & Utilities
ipython==8.14.0
jupyter==1.0.0
